not include payment

+++ FahasaStoreClientApp/Entities/Dimension.cs
@@ -11,9 +11,9 @@ namespace FahasaStoreClientApp.Entities
         }
 
         public int Id { get; set; }
-        public double Length { get; set; }
-        public double Width { get; set; }
-        public double Height { get; set; }
+        public int Length { get; set; }
+        public int Width { get; set; }
+        public int Height { get; set; }
         public string Unit { get; set; } = null!;
         public DateTime? CreatedAt { get; set; }
